# datalith/src/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Datalith API",
    "version": "1.0.0",
    "description": "A media store that runs as one instance, with file contents on disk and metadata in SQLite. IDs are UUID strings. Dates are UTC RFC 3339 strings. File sizes and page totals are decimal strings. A trusted backend or reverse proxy must handle authentication. This release does not include audio or video conversion or a new Node.js client.",
    "license": {
      "name": "MIT",
      "identifier": "MIT"
    }
  },
  "servers": [
    {
      "url": "/api/v1"
    }
  ],
  "security": [],
  "paths": {
    "/uploads": {
      "post": {
        "operationId": "upload",
        "summary": "Save an upload and create a task",
        "responses": {
          "202": {
            "description": "Save an upload and create a task",
            "headers": {
              "X-Request-Id": {
                "schema": {
                  "type": "string",
                  "format": "uuid"
                },
                "description": "Request ID created by the server."
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Task"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "404": {
            "$ref": "#/components/responses/Error"
          },
          "409": {
            "$ref": "#/components/responses/Error"
          },
          "413": {
            "$ref": "#/components/responses/Error"
          },
          "415": {
            "$ref": "#/components/responses/Error"
          },
          "422": {
            "$ref": "#/components/responses/Error"
          },
          "500": {
            "$ref": "#/components/responses/Error"
          },
          "503": {
            "$ref": "#/components/responses/Error"
          }
        },
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 128
            },
            "description": "An ASCII key stored with the task for seven days by default. The same key and input return the existing task. Different input with the same key returns 409."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": [
                  "file"
                ],
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary"
                  },
                  "options": {
                    "$ref": "#/components/schemas/UploadOptions"
                  }
                }
              },
              "encoding": {
                "options": {
                  "contentType": "application/json"
                }
              }
            }
          }
        },
        "description": "The service saves the upload and task before it responds. It does not wait for image conversion. Without options, it creates a permanent resource. The upload is streamed to disk."
      }
    },
    "/tasks/{id}": {
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string",
            "format": "uuid"
          }
        }
      ],
      "get": {
        "operationId": "getTask",
        "summary": "Read task state and result",
        "responses": {
          "200": {
            "description": "Read task state and result",
            "headers": {
              "X-Request-Id": {
                "schema": {
                  "type": "string",
                  "format": "uuid"
                },
                "description": "Request ID created by the server."
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Task"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "404": {
            "$ref": "#/components/responses/Error"
          },
          "409": {
            "$ref": "#/components/responses/Error"
          },
          "413": {
            "$ref": "#/components/responses/Error"
          },
          "415": {
            "$ref": "#/components/responses/Error"
          },
          "422": {
            "$ref": "#/components/responses/Error"
          },
          "500": {
            "$ref": "#/components/responses/Error"
          },
          "503": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
    "/tasks/{id}/cancel": {
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string",
            "format": "uuid"
          }
        }
      ],
      "post": {
        "operationId": "cancelTask",
        "summary": "Ask a task to stop between processing steps",
        "responses": {
          "202": {
            "description": "Ask a task to stop between processing steps",
            "headers": {
              "X-Request-Id": {
                "schema": {
                  "type": "string",
                  "format": "uuid"
                },
                "description": "Request ID created by the server."
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Task"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "404": {
            "$ref": "#/components/responses/Error"
          },
          "409": {
            "$ref": "#/components/responses/Error"
          },
          "413": {
            "$ref": "#/components/responses/Error"
          },
          "415": {
            "$ref": "#/components/responses/Error"
          },
          "422": {
            "$ref": "#/components/responses/Error"
          },
          "500": {
            "$ref": "#/components/responses/Error"
          },
          "503": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
    "/tasks/{id}/retry": {
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string",
            "format": "uuid"
          }
        }
      ],
      "post": {
        "operationId": "retryTask",
        "summary": "Retry a failed or cancelled task",
        "responses": {
          "202": {
            "description": "Retry a failed or cancelled task",
            "headers": {
              "X-Request-Id": {
                "schema": {
                  "type": "string",
                  "format": "uuid"
                },
                "description": "Request ID created by the server."
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Task"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "404": {
            "$ref": "#/components/responses/Error"
          },
          "409": {
            "$ref": "#/components/responses/Error"
          },
          "413": {
            "$ref": "#/components/responses/Error"
          },
          "415": {
            "$ref": "#/components/responses/Error"
          },
          "422": {
            "$ref": "#/components/responses/Error"
          },
          "500": {
            "$ref": "#/components/responses/Error"
          },
          "503": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
...
    "/media/{id}": {
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string",
            "format": "uuid"
          }
        }
      ],
      "get": {
        "operationId": "getMedia",
        "summary": "Read metadata without using a single-use download claim",
        "responses": {
          "200": {
            "description": "Read metadata without using a single-use download claim",
            "headers": {
              "X-Request-Id": {
                "schema": {
                  "type": "string",
                  "format": "uuid"
                },
                "description": "Request ID created by the server."
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Media"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "404": {
            "$ref": "#/components/responses/Error"
          },
          "409": {
            "$ref": "#/components/responses/Error"
          },
          "413": {
            "$ref": "#/components/responses/Error"
          },
          "415": {
            "$ref": "#/components/responses/Error"
          },
          "422": {
            "$ref": "#/components/responses/Error"
          },
          "500": {
            "$ref": "#/components/responses/Error"
          },
          "503": {
            "$ref": "#/components/responses/Error"
          }
        }
      },
      "delete": {
        "operationId": "deleteMedia",
        "summary": "Delete media and clean up unused files later",
        "responses": {
          "204": {
            "description": "Delete media and clean up unused files later",
            "headers": {
              "X-Request-Id": {
                "schema": {
                  "type": "string",
                  "format": "uuid"
                },
                "description": "Request ID created by the server."
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "404": {
            "$ref": "#/components/responses/Error"
          },
          "409": {
            "$ref": "#/components/responses/Error"
          },
          "413": {
            "$ref": "#/components/responses/Error"
          },
          "415": {
            "$ref": "#/components/responses/Error"
          },
          "422": {
            "$ref": "#/components/responses/Error"
          },
          "500": {
            "$ref": "#/components/responses/Error"
          },
          "503": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
    "/media/{id}/tasks": {
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string",
            "format": "uuid"
          }
        }
      ],
      "post": {
        "operationId": "processMedia",
        "summary": "Create an image task from a saved original",
        "responses": {
          "202": {
            "description": "Create an image task from a saved original",
            "headers": {
              "X-Request-Id": {
                "schema": {
                  "type": "string",
                  "format": "uuid"
                },
                "description": "Request ID created by the server."
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Task"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "404": {
            "$ref": "#/components/responses/Error"
          },
          "409": {
            "$ref": "#/components/responses/Error"
          },
          "413": {
            "$ref": "#/components/responses/Error"
          },
          "415": {
            "$ref": "#/components/responses/Error"
          },
          "422": {
            "$ref": "#/components/responses/Error"
          },
          "500": {
            "$ref": "#/components/responses/Error"
          },
          "503": {
            "$ref": "#/components/responses/Error"
          }
        },
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 128
            },
            "description": "An ASCII key stored with the task for seven days by default. The same key and input return the existing task. Different input with the same key returns 409."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProcessOptions"
              }
            }
          }
        },
        "description": "The original file must be available. A single-use source returns 409. The result keeps the source expiry time and is not saved if that time has passed."
      }
    },
    "/exports": {
      "post": {
        "operationId": "exportMedia",
        "summary": "Export all media or selected IDs",
        "responses": {
          "202": {
            "description": "Export all media or selected IDs",
            "headers": {
              "X-Request-Id": {
                "schema": {
                  "type": "string",
                  "format": "uuid"
                },
                "description": "Request ID created by the server."
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Task"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "404": {
            "$ref": "#/components/responses/Error"
          },
          "409": {
            "$ref": "#/components/responses/Error"
          },
          "413": {
            "$ref": "#/components/responses/Error"
          },
          "415": {
            "$ref": "#/components/responses/Error"
          },
          "422": {
            "$ref": "#/components/responses/Error"
          },
          "500": {
            "$ref": "#/components/responses/Error"
          },
          "503": {
            "$ref": "#/components/responses/Error"
          }
        },
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 128
            },
            "description": "An ASCII key stored with the task for seven days by default. The same key and input return the existing task. Different input with the same key returns 409."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExportOptions"
              }
            }
          }
        },
        "description": "Export pauses content writes, file cleanup, and single-use downloads. Normal downloads continue. The archive contains a versioned manifest and one copy of each file content."
      }
    },
    "/imports": {
      "post": {
        "operationId": "importMedia",
        "summary": "Check and import a Datalith archive",
        "responses": {
          "202": {
            "description": "Check and import a Datalith archive",
            "headers": {
              "X-Request-Id": {
                "schema": {
                  "type": "string",
                  "format": "uuid"
                },
                "description": "Request ID created by the server."
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Task"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "404": {
            "$ref": "#/components/responses/Error"
          },
          "409": {
            "$ref": "#/components/responses/Error"
          },
          "413": {
            "$ref": "#/components/responses/Error"
          },
          "415": {
            "$ref": "#/components/responses/Error"
          },
          "422": {
            "$ref": "#/components/responses/Error"
          },
          "500": {
            "$ref": "#/components/responses/Error"
          },
          "503": {
            "$ref": "#/components/responses/Error"
          }
        },
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 128
            },
            "description": "An ASCII key stored with the task for seven days by default. The same key and input return the existing task. Different input with the same key returns 409."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": [
                  "file"
                ],
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              }
            }
          }
        },
        "description": "Import keeps IDs where possible and skips matching items. An ID used by different data is replaced with a new ID. The result includes the old-to-new ID maps. Importing the same archive again does not create duplicates."
      }
    },
...
    "/capabilities": {
      "get": {
        "operationId": "getCapabilities",
        "summary": "Read supported media types, limits, and planned output formats",
        "responses": {
          "200": {
            "description": "Read supported media types, limits, and planned output formats",
            "headers": {
              "X-Request-Id": {
                "schema": {
                  "type": "string",
                  "format": "uuid"
                },
                "description": "Request ID created by the server."
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Capabilities"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "404": {
            "$ref": "#/components/responses/Error"
          },
          "409": {
            "$ref": "#/components/responses/Error"
          },
          "413": {
            "$ref": "#/components/responses/Error"
          },
          "415": {
            "$ref": "#/components/responses/Error"
          },
          "422": {
            "$ref": "#/components/responses/Error"
          },
          "500": {
            "$ref": "#/components/responses/Error"
          },
          "503": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string"
              },
              "message": {
                "type": "string"
              }
            },
            "required": [
              "code",
              "message"
            ]
          },
          "request_id": {
            "type": "string",
            "format": "uuid"
          }
        },
        "required": [
          "error",
          "request_id"
        ]
      },
      "MediaKind": {
        "type": "string",
        "enum": [
          "resource",
          "image",
          "audio",
          "video"
        ],
        "description": "Audio and video are planned for a future release. Both currently return 415."
      },
...
      "UploadOptions": {
        "type": "object",
        "properties": {
          "kind": {
            "allOf": [
              {
                "$ref": "#/components/schemas/MediaKind"
              }
            ],
            "default": "resource"
          },
          "file_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          },
          "file_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          },
          "retention": {
            "$ref": "#/components/schemas/Retention"
          },
          "image": {
            "$ref": "#/components/schemas/ImageOptions"
          }
        },
        "additionalProperties": false
      },
      "ProcessOptions": {
        "type": "object",
        "properties": {
          "image": {
            "$ref": "#/components/schemas/ImageOptions"
          }
        },
        "additionalProperties": false,
        "description": "The original file must be available. A single-use source returns 409. The result keeps the source expiry time and is not saved if that time has passed."
      },
      "ExportOptions": {
        "type": "object",
        "properties": {
          "ids": {
            "anyOf": [
              {
                "type": "array",
                "items": {
                  "type": "string",
                  "format": "uuid"
                },
                "minItems": 1,
                "maxItems": 100000
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "additionalProperties": false
      },
...
      "Media": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "kind": {
            "$ref": "#/components/schemas/MediaKind"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "file_name": {
            "type": "string"
          },
          "original": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/MediaFile"
              },
              {
                "type": "null"
              }
            ]
          },
          "variants": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Variant"
            }
          },
          "expires_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ]
          },
          "single_use": {
            "type": "boolean"
          },
          "consumed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ]
          },
          "animated": {
            "type": "boolean"
          },
          "frame_count": {
            "type": "integer",
            "minimum": 0
          }
        },
        "required": [
          "id",
          "kind",
          "created_at",
          "file_name",
          "original",
          "variants",
          "expires_at",
          "single_use",
          "consumed_at",
          "animated",
          "frame_count"
        ]
      },
...
      "TaskError": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ]
      },
      "ExportResult": {
        "type": "object",
        "properties": {
          "artifact_path": {
            "type": "string"
          },
          "media_count": {
            "type": "integer",
            "minimum": 0
          },
          "artifact": {
            "$ref": "#/components/schemas/MediaFile"
          }
        },
        "required": [
          "artifact_path",
          "media_count",
          "artifact"
        ]
      },
      "ImportResult": {
        "type": "object",
        "properties": {
          "imported": {
            "type": "integer",
            "minimum": 0
          },
          "skipped": {
            "type": "integer",
            "minimum": 0
          },
          "id_map": {
            "type": "object",
            "additionalProperties": {
              "type": "string",
              "format": "uuid"
            }
          },
          "archive_id": {
            "type": "string",
            "format": "uuid"
          },
          "file_id_map": {
            "type": "object",
            "additionalProperties": {
              "type": "string",
              "format": "uuid"
            }
          }
        },
        "required": [
          "imported",
          "skipped",
          "id_map",
          "archive_id",
          "file_id_map"
        ]
      },
      "Task": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "kind": {
            "type": "string",
            "enum": [
              "resource",
              "image",
              "import",
              "export"
            ]
          },
          "status": {
            "type": "string",
            "enum": [
              "queued",
              "running",
              "cancelling",
              "succeeded",
              "failed",
              "cancelled"
            ]
          },
          "stage": {
            "type": "string"
          },
          "completed_units": {
            "type": "integer",
            "minimum": 0
          },
          "total_units": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0
              },
              {
                "type": "null"
              }
            ]
          },
          "attempt": {
            "type": "integer",
            "minimum": 0,
            "description": "Number of times processing has started. This is zero before the first attempt."
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          },
          "result": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/Media"
              },
              {
                "$ref": "#/components/schemas/ExportResult"
              },
              {
                "$ref": "#/components/schemas/ImportResult"
              },
              {
                "type": "null"
              }
            ],
            "description": "Upload and image tasks return Media. Export tasks return ExportResult. Import tasks return ImportResult."
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TaskError"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "id",
          "kind",
          "status",
          "stage",
          "completed_units",
          "total_units",
          "attempt",
          "created_at",
          "updated_at",
          "result",
          "error"
        ]
      },
...
    },
    "responses": {
      "Error": {
        "description": "An error object with a code and message. Writes paused by an export return 503 with Retry-After: 1.",
        "headers": {
          "X-Request-Id": {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "Request ID created by the server."
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      }
    }
  }
}